StudentEnrollment to Student keeping Student Enrollment model and updating test Adding .XML comments to Authenticate Controller
