+++ Lms/LMS.UnitTests/StudentControllerTests.cs
@@ -17,7 +17,7 @@ namespace LMS.UnitTests
 {
     #nullable disable warnings
     [TestClass]
-    public class StudentControllerTests
+    public class StudentEnrollmentControllerTests
     {
         Mock<IStudentDao> mockStudentDao;
         StudentController sut;
