+++ Lms/LMS.UnitTests/StudentEnrollmentControllerTests.cs
@@ -15,12 +15,12 @@ using System;
 
 namespace LMS.UnitTests
 {
-    #nullable disable warnings
+#nullable disable warnings
     [TestClass]
-    public class StudentEnrollmentControllerTests
+    public class StudentControllerTests
     {
-        private Mock<IStudentEnrollmentDao> _mockStudentEnrollmentDao;
-        private StudentEnrollmentController _sut;
+        private Mock<IStudentDao> _mockStudentDao;
+        private StudentController _sut;
         private Guid _studentGuid;
         private Guid _invalidStudentGuid;
         private List<StudentEnrollmentModel> _studentEnrollment;
@@ -29,9 +29,9 @@ namespace LMS.UnitTests
         [TestInitialize]
         public void Initialize()
         {
-            _mockStudentEnrollmentDao = new Mock<IStudentEnrollmentDao>();
+            _mockStudentDao = new Mock<IStudentDao>();
             _cache = new MemoryCache(new MemoryCacheOptions());
-            _sut = new StudentEnrollmentController(_mockStudentEnrollmentDao.Object, _cache);
+            _sut = new StudentController(_mockStudentDao.Object, _cache);
             _studentGuid = new Guid("0AE43554-0BB1-42B1-94C7-04420A2167A9");
             _invalidStudentGuid = new Guid("00000000-0000-0000-0000-000000000000");
 
@@ -53,21 +53,21 @@ namespace LMS.UnitTests
         [TestCleanup]
         public void Cleanup()
         {
-            _mockStudentEnrollmentDao = null;
+            _mockStudentDao = null;
             _sut = null;
             _studentGuid = new Guid();
             _studentEnrollment = null;
         }
 
         [TestMethod]
-        public async Task GetStudentEnrollment_ReturnsOkResponse_WhenStudentIdFoundInDatabase()
+        public async Task GetStudent_ReturnsOkResponse_WhenStudentIdFoundInDatabase()
         {
             // Arrange
-            _mockStudentEnrollmentDao.Setup(x => x.GetStudentEnrollmentHistory(_studentGuid, null, null, null,null))
+            _mockStudentDao.Setup(x => x.GetStudentEnrollmentHistory(_studentGuid, null, null, null))
                 .ReturnsAsync(_studentEnrollment);
 
             // Act
-            var result = await _sut.GetStudentEnrollmentHistory(_studentGuid, null, null, null,null);
+            var result = await _sut.GetStudentEnrollmentHistory(_studentGuid, null, null, null);
 
             // Assert
             var okResult = result as OkObjectResult;
@@ -81,18 +81,18 @@ namespace LMS.UnitTests
             Assert.IsNotNull(studentResult);
             CollectionAssert.AreEqual(_studentEnrollment, studentResult);
 
-           _mockStudentEnrollmentDao.Verify(x => x.GetStudentEnrollmentHistory(_studentGuid, null, null, null, null), Times.Once);
+            _mockStudentDao.Verify(x => x.GetStudentEnrollmentHistory(_studentGuid, null, null, null), Times.Once);
         }
 
         [TestMethod]
-        public async Task GetStudentEnrollment_ReturnsNotFoundResponse_WhenStudentGuidIsInvalid()
+        public async Task GetStudent_ReturnsNotFoundResponse_WhenStudentGuidIsInvalid()
         {
             // Arrange
-            _mockStudentEnrollmentDao.Setup(x => x.GetStudentEnrollmentHistory(_invalidStudentGuid, null, null, null, null))
+            _mockStudentDao.Setup(x => x.GetStudentEnrollmentHistory(_invalidStudentGuid, null, null, null))
                 .ReturnsAsync((IEnumerable<StudentEnrollmentModel>)null);
 
             // Act
-            var result = await _sut.GetStudentEnrollmentHistory(_invalidStudentGuid, null, null, null, null);
+            var result = await _sut.GetStudentEnrollmentHistory(_invalidStudentGuid, null, null, null);
             var notFoundResult = result as NotFoundObjectResult;
             var apiResponse = notFoundResult.Value as ApiResponse;
