+++ Lms/Lms/Daos/IStudentDao.cs
@@ -9,7 +9,10 @@ namespace Lms.Daos
     {
         Task CreateStudent(StudentModel newStudent);
         Task<T> GetStudentById<T>(Guid id);
+        Task<IEnumerable<StudentEnrollmentModel>> GetStudentEnrollmentHistory(Guid StudentId, string StudentPhone, string Cancelled, string HasPassed);
+        Task<IEnumerable<StudentEnrollmentModel>> GetStudentsInCourseByCourseId(Guid id);
         Task PartiallyUpdateStudentById(StudentModel updateRequest);
-        Task DeleteStudentById(Guid id);      
+        Task DeleteStudentById(Guid id);
+        
     }
 }
