+++ Lms/Lms/Daos/StudentDao.cs
@@ -1,8 +1,11 @@
 ﻿using Dapper;
 using Lms.Models;
 using Lms.Wrappers;
+using Microsoft.Identity.Client;
 using System;
+using System.Collections.Generic;
 using System.Data;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Lms.Daos
@@ -50,6 +53,76 @@ namespace Lms.Daos
                 return student;
             }
         }
+
+        //Get Student Enrollment by Parameters
+        public async Task<IEnumerable<StudentEnrollmentModel>> GetStudentEnrollmentHistory(Guid StudentId, string StudentPhone = null, string Cancelled = null, string HasPassed = null)
+        {
+            var query = $"SELECT" +
+            $"  C.[CourseId]" +
+            $", C.[CourseName]" +
+            $", C.[CourseStatus]" +
+            $", C.[StartDate]" +
+            $", C.[EndDate]" +
+            $", SE.[Cancelled]" +
+            $", SE.[CancellationReason]" +
+            $", SE.[HasPassed]" +
+            $", T.[TeacherEmail]" +
+            $", S.[StudentEmail]" +
+            $" FROM [StudentEnrollmentLog] as SE" +
+            $" INNER JOIN [Course] as C ON SE.[CourseId] = C.[CourseId]" +
+            $" INNER JOIN [Teacher] as T ON C.[TeacherId] = T.[TeacherId]" +
+            $" INNER JOIN [Student] as S ON SE.[StudentId] = S.[StudentId]" +
+            $" WHERE 1=1";
+
+            var parameters = new DynamicParameters();
+            if (StudentId != Guid.Empty)
+            {
+                query += " AND S.StudentId = @StudentId";
+                parameters.Add("StudentId", StudentId, DbType.Guid);
+            }
+            if (!string.IsNullOrEmpty(StudentPhone))
+            {
+                query += " AND StudentPhone = @StudentPhone";
+                parameters.Add("StudentPhone", StudentPhone, DbType.String);
+            }
+            if (!string.IsNullOrEmpty(Cancelled))
+            {
+                query += " AND Cancelled = @Cancelled";
+                parameters.Add("Cancelled", Cancelled, DbType.String);
+            }
+            if (!string.IsNullOrEmpty(HasPassed))
+            {
+                query += " OR HasPassed = @HasPassed";
+                parameters.Add("HasPassed", HasPassed, DbType.String);
+            }
+
+            using (sqlWrapper.CreateConnection())
+            {
+                var studentHistory = await sqlWrapper.QueryAsync<StudentEnrollmentModel>(query, parameters);
+
+                return studentHistory.ToList();
+            }
+        }
+
+        //Get Student Enrollment by CourseId
+        public async Task<IEnumerable<StudentEnrollmentModel>> GetStudentsInCourseByCourseId(Guid courseId)
+        {
+            var query = $"SELECT * " +
+            $"FROM [StudentEnrollmentLog] " +
+            $"INNER JOIN [Student] ON [Student].[StudentId] = [StudentEnrollmentLog].[StudentId] " +
+            $"WHERE CourseId = @CourseId " +
+            $"ORDER BY StudentLastName";
+
+            var parameters = new DynamicParameters();
+            parameters.Add("CourseId", courseId, DbType.Guid);
+
+            using (sqlWrapper.CreateConnection())
+            {
+                var course = await sqlWrapper.QueryAsync<StudentEnrollmentModel>(query, parameters);
+                return course;
+            }
+        }
+    
 
         // PATCH a student within the Student table. 
         public async Task PartiallyUpdateStudentById(StudentModel updateRequest)
